+++ BusFast/Scrape/Stop.cs
@@ -2,7 +2,7 @@
 {
     public class Stop
     {
-        public int Stop_Id { get; set; }
+        public long Stop_Id { get; set; }
         public string Stop_Name { get; set; }
     }
 }
